+++ src/CSharp.Data.Adaptive/ASet.fs
@@ -47,6 +47,17 @@ type AdaptiveHashSet private() =
 
     [<MethodImpl(MethodImplOptions.AggressiveInlining)>]
     static member OfHashSet<'T>(set: HashSet<'T>) = ASet.ofHashSet set
+
+    /// Creates an AdaptiveHashSet from a tree of lists
+    /// NOTE: does not expect duplicates -> TODO
+    [<MethodImpl(MethodImplOptions.AggressiveInlining)>]
+    static member OfListTree(list: alist<'TNode>, getChildren : Func<'TNode, alist<'TNode>>) =
+        list |> ASet.ofListTree getChildren.Invoke
+
+    /// Creates an AdaptiveHashSet from a tree of sets
+    [<MethodImpl(MethodImplOptions.AggressiveInlining)>]
+    static member OfSetTree(set: aset<'TNode>, getChildren : Func<'TNode, aset<'TNode>>) =
+        set |> ASet.ofSetTree getChildren.Invoke
         
     [<Extension; MethodImpl(MethodImplOptions.AggressiveInlining)>]
     static member ToAdaptiveHashSet(this: seq<'T>) = ASet.ofSeq this
@@ -272,14 +283,13 @@ type AdaptiveHashSet private() =
         this |> ASet.toAList
         
     [<Extension; MethodImpl(MethodImplOptions.AggressiveInlining)>]
-    static member MapToMap(this: aset<'T>, mapping: Func<'T, 'Value>) =
-        this |> ASet.mapToAMap mapping.Invoke
+    static member MapToMap(this: aset<'T>, getValue: Func<'T, 'Value>) =
+        this |> ASet.mapToAMap getValue.Invoke
         
     [<Extension; MethodImpl(MethodImplOptions.AggressiveInlining)>]
-    static member GroupBy(this: aset<'T>, mapping: Func<'T, 'G>) =
-        this |> ASet.groupBy mapping.Invoke
-
-        
+    static member GroupBy(this: aset<'T>, getKey: Func<'T, 'G>) =
+        this |> ASet.groupBy getKey.Invoke
+                
     [<Extension; MethodImpl(MethodImplOptions.AggressiveInlining)>]
     static member ToAdaptiveHashMap(this: aset<'K * 'V>) =
         AMap.ofASet this
@@ -288,6 +298,9 @@ type AdaptiveHashSet private() =
     static member ToAdaptiveHashMapIgnoreDuplicates(this: aset<'K * 'V>) =
         AMap.ofASetIgnoreDuplicates this
 
+    [<Extension; MethodImpl(MethodImplOptions.AggressiveInlining)>]
+    static member ToAdaptiveHashMapIgnoreDuplicates(this: aset<'TValue>, getKey : Func<'TValue, 'TKey>) =
+        this |> AMap.ofASetMappedIgnoreDuplicates getKey.Invoke
 
     [<Extension; MethodImpl(MethodImplOptions.AggressiveInlining)>]
     static member ToArray(this: aset<'T>) =