+++ src/FSharp.Data.Adaptive/Datastructures/HashMapDelta.fs
@@ -77,6 +77,12 @@ module HashMapDelta =
     /// A HashMapDelta containing all the given deltas.
     let ofSeq (elements : seq<'K * ElementOperation<'V>>) = 
         HashMapDelta(HashMap.ofSeq elements)
+
+    #if !FABLE_COMPILER
+    /// A HashMapDelta containing all the given deltas.
+    let ofSeqV (elements : seq<struct('K * ElementOperation<'V>)>) = 
+        HashMapDelta(HashMap.ofSeqV elements)
+    #endif
         
     /// A HashMapDelta containing all the given deltas.
     let ofList (elements : list<'K * ElementOperation<'V>>) =